app/requirements.txt
# ==============================================================================
# FastAPI Task Manager - Dependencies
# ==============================================================================
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
sqlalchemy[asyncio]>=2.0.0,<3.0.0
asyncpg>=0.29.0,<1.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
slowapi>=0.1.9,<1.0.0
prometheus-fastapi-instrumentator>=7.0.0,<8.0.0
structlog>=24.0.0,<25.0.0
alembic>=1.13.0,<2.0.0
